+++ flake.nix
@@ -2,7 +2,7 @@
   description = "Musium";
 
   inputs.nixpkgs.url = "nixpkgs/nixos-23.05";
-  inputs.squiller.url = "github:ruuda/squiller?ref=v0.3.0";
+  inputs.squiller.url = "github:ruuda/squiller?ref=v0.4.0";
   inputs.squiller.inputs.nixpkgs.follows = "nixpkgs";
 
   outputs = { self, nixpkgs, squiller }: