+++ src/database.sql
@@ -58,6 +58,25 @@ create table if not exists listens
 create unique index if not exists ix_listens_unique_second
 on listens (cast(strftime('%s', started_at) as integer));
 
+create table if not exists ratings
+( id          integer primary key
+-- ISO-8601 time with UTC offset at which we rated the track.
+, created_at  string  not null unique
+-- Musium track that we are rating. We don't enforce a foreign key relation
+-- here, such that when we re-import a track we don't lose the rating data. The
+-- downside is that we may end up with dangling ratings if tracks get deleted
+-- or moved (e.g. a correction in track number), but that's acceptable.
+, track_id    integer not null
+-- The rating for this track.
+, rating      integer not null check ((rating >= -1) and (rating <= 2))
+-- "musium" for ratings created from Musium, otherwise the source that the
+-- rating was imported from, e.g. "last.fm".
+, source      string not null
+);
+
+create unique index if not exists ix_ratings_unique_second
+on ratings (cast(strftime('%s', created_at) as integer));
+
 create table if not exists files
 -- First an id, and properties about the file, but not its contents.
 -- We can use this to see if a file needs to be re-scanned. The mtime
@@ -292,3 +311,35 @@ from
   listens
 group by
   album_id;
+
+-- Insert a rating for a given track.
+--
+-- When the `created_at` timestamp is not unique, this replaces the previous
+-- rating that was present for that timestamp. This might happen when the user
+-- edits the rating in quick succession; then we only store the last write.
+-- @query insert_or_replace_rating(track_id: i64, created_at: str, rating: i64)
+insert or replace into
+  ratings (track_id, created_at, rating, source)
+values
+  (:track_id, :created_at, :rating, 'musium');
+
+-- Backfill a rating for a given track.
+--
+-- The timestamp must be unique on the second.
+-- @query insert_rating(track_id: i64, created_at: str, rating: i64, source: str)
+insert into
+  ratings (track_id, created_at, rating, source)
+values
+  (:track_id, :created_at, :rating, :source);
+
+-- @query iter_ratings() ->* TrackRating
+select
+    id       -- :i64
+  , track_id -- :i64
+  , rating   -- :i64
+from
+  ratings
+order by
+  -- Order by ascending creation time to ensure we can clamp to rating ranges,
+  -- should we need to. We have an index on this expression.
+  cast(strftime('%s', created_at) as integer) asc;