+++ src/history.rs
@@ -9,6 +9,9 @@
 
 use std::path::Path;
 use std::sync::mpsc::Receiver;
+use std::sync::{Arc, Mutex};
+
+use chrono::{SecondsFormat, Utc};
 
 use crate::database_utils;
 use crate::database as db;
@@ -16,18 +19,26 @@ use crate::database::{Connection, Listen, Result};
 use crate::mvar::Var;
 use crate::player::QueueId;
 use crate::{MetaIndex, MemoryMetaIndex, TrackId};
+use crate::user_data::{Rating, UserData};
 
-/// Changes in the playback state to be recorded.
+/// Changes in the playback state or library to be recorded.
 pub enum PlaybackEvent {
     Started(QueueId, TrackId),
     Completed(QueueId, TrackId),
     QueueEnded,
+
+    /// The user modified the rating for the given track.
+    Rated {
+        track_id: TrackId,
+        rating: Rating,
+    },
 }
 
 /// Main for the thread that logs historical playback events.
 pub fn main(
     db_path: &Path,
     index_var: Var<MemoryMetaIndex>,
+    user_data: Arc<Mutex<UserData>>,
     events: Receiver<PlaybackEvent>,
 ) -> Result<()> {
     let connection = database_utils::connect_read_write(db_path)?;
@@ -36,9 +47,9 @@ pub fn main(
     let mut last_listen_id = None;
 
     for event in events {
-        let now = chrono::Utc::now();
+        let now = Utc::now();
         let use_zulu_suffix = true;
-        let now_str = now.to_rfc3339_opts(chrono::SecondsFormat::Millis, use_zulu_suffix);
+        let now_str = now.to_rfc3339_opts(SecondsFormat::Millis, use_zulu_suffix);
 
         match event {
             PlaybackEvent::Started(queue_id, track_id) => {
@@ -91,8 +102,19 @@ pub fn main(
                 // needed, but I back up my database with rsync once in a
                 // while, and I like to have everything in one file instead
                 // of having to sync the WAL as well. We checkpoint after
-                // the queue ends, before after the post-playback program runs.
+                // the queue ends, before the post-playback program runs.
                 connection.execute("PRAGMA wal_checkpoint(PASSIVE);")?;
+            }
+            PlaybackEvent::Rated { track_id, rating } => {
+                let mut tx = db.begin()?;
+                db::insert_or_replace_rating(
+                    &mut tx,
+                    track_id.0 as i64,
+                    &now_str,
+                    rating as i64,
+                )?;
+                tx.commit()?;
+                user_data.lock().unwrap().set_track_rating(track_id, rating);
             }
         }
     }